% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continous-scale.R
\name{scale_colour_ymy_seq_c}
\alias{scale_colour_ymy_seq_c}
\title{Continuous colour scale from YmyColour}
\usage{
scale_colour_ymy_seq_c(
  name = names(ymy_colours)[1],
  direction = 1,
  amount = 0.2,
  ...
)
}
\arguments{
\item{name}{character string, specifying the name of the colour to be used. Colours in YmyColour package include orange, blue, green, coral, yellow and rust.}

\item{direction}{Sets the order of colours in the scale. If 1, the default, colours are ordered as default order. If -1, the order of colours is reversed.}

\item{amount}{numeric specifying the amount of lightening. This is applied either multiplicatively or additively to the luminance value,
depending on the setting of method (either relative or absolute). Negative numbers cause darkening.}

\item{...}{Other arguments passed on to \code{\link[ggplot2]{discrete_scale}}, \code{\link[ggplot2]{continuous_scale}},
or \code{\link[ggplot2]{binned_scale}} to control name, limits, breaks, labels and so forth.}
}
\description{
Continuous colour scale from YmyColour
}
\details{
The ymy scales were carefully designed and tested on continuous data.
}
\examples{
# Set the amount of lightening
df <- data.frame(x = 1:6, y = 1:6, type = letters[1:6])
ggplot(df, aes(x = x, y = y,colour = y)) +
geom_point(size = 10) +
theme_bw() +
scale_colour_ymy_seq_c(amount = 0.6)

# Set reversed order for gradient colour
df <- data.frame(x = 1:6, y = 1:6, type = letters[1:6])
ggplot(df, aes(x = x, y = y,colour = y)) +
geom_point(size = 10) +
theme_bw() +
scale_colour_ymy_seq_c(amount = 0.6, direction = -1)
}
\seealso{
Other colour scales: \code{\link[ggplot2]{scale_colour_brewer}}
}
