% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{ymy_palette}
\alias{ymy_palette}
\title{Ymy Colour Palette}
\usage{
ymy_palette(palette = "ymy_colours", direction = 1)
}
\arguments{
\item{palette}{character string, specifying the name of the palette to be used.}

\item{direction}{numeric, specifying the direction of the palette to be used.}
}
\value{
A character vector of hex colour codes of 6.
}
\description{
Ymy Colour Palette
}
\details{
The ymy colour palette has 6 colours and the function is designed to allow users to choose the number of colours to be used and the direction of the colours to be applied.
}
\examples{
ymy_palette()(2)
ymy_palette(direction = -1)(3)
}
