% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{ymy_palette}
\alias{ymy_palette}
\title{Ymy Colour Palette}
\usage{
ymy_palette(palette = "ymy_colours", direction = 1)
}
\arguments{
\item{palette}{character string, specifying the name of the palette to be used}

\item{direction}{numeric, specifying the direction of the palette to be used}

\item{n}{numeric, specifying the number of colours to be used}
}
\value{
Colours in the palette
}
\description{
Ymy Colour Palette
}
\details{
This function is designed to allow users to choose the number of colours to be used and the direction of the colours to be applied
}
\examples{
ymy_palette()(2)
ymy_palette(direction = -1)(3)
}
